Move local utils to utils package
