+++ services/streaming_service.py
@@ -2,7 +2,7 @@
 from datetime import datetime, timezone
 
 try:  # pragma: no cover - prefer local stub if available
-    import aiosqlite_local as aiosqlite
+    import utils.aiosqlite_local as aiosqlite
 except ModuleNotFoundError:  # pragma: no cover - fallback to package
     import aiosqlite  # type: ignore
 from backend.database import DB_PATH
